Reuse color and daterange components
